Warnings in OpenBK7231T_App fixed. (#1098)
* Warnings in AppT fixed.

* Remove ota include file added by mistake in drv_sm2235.

+++ src/driver/drv_tuyaMCU.c
@@ -470,7 +470,7 @@ void TuyaMCU_SendStateInternal(uint8_t id, uint8_t type, void* value, int dataLe
 {
 	uint16_t payload_len = 0;
 	
-	payload_len = TuyaMCU_AppendStateInternal(g_tuyaMCUpayloadBuffer, g_tuyaMCUpayloadBuffer,
+	payload_len = TuyaMCU_AppendStateInternal(g_tuyaMCUpayloadBuffer, (int)g_tuyaMCUpayloadBuffer,
 		payload_len, id, type, value, dataLen);
 
 	TuyaMCU_SendCommandWithData(TUYA_CMD_SET_DP, g_tuyaMCUpayloadBuffer, payload_len);
@@ -484,13 +484,13 @@ void TuyaMCU_SendTwoVals(byte idA, int valA, byte idB, int valB)
 	swap[1] = ((byte*)&valA)[2];
 	swap[2] = ((byte*)&valA)[1];
 	swap[3] = ((byte*)&valA)[0];
-	payload_len = TuyaMCU_AppendStateInternal(g_tuyaMCUpayloadBuffer, g_tuyaMCUpayloadBuffer,
+	payload_len = TuyaMCU_AppendStateInternal(g_tuyaMCUpayloadBuffer, (int)g_tuyaMCUpayloadBuffer,
 		payload_len, idA, DP_TYPE_VALUE, swap, 4);
 	swap[0] = ((byte*)&valB)[3];
 	swap[1] = ((byte*)&valB)[2];
 	swap[2] = ((byte*)&valB)[1];
 	swap[3] = ((byte*)&valB)[0];
-	payload_len = TuyaMCU_AppendStateInternal(g_tuyaMCUpayloadBuffer, g_tuyaMCUpayloadBuffer,
+	payload_len = TuyaMCU_AppendStateInternal(g_tuyaMCUpayloadBuffer, (int)g_tuyaMCUpayloadBuffer,
 		payload_len, idB, DP_TYPE_VALUE, swap, 4);
 
 	TuyaMCU_SendCommandWithData(TUYA_CMD_SET_DP, g_tuyaMCUpayloadBuffer, payload_len);
@@ -976,7 +976,7 @@ void TuyaMCU_SendStateRawFromString(int dpId, const char *args) {
 }
 const char *STR_FindArg(const char *s, int arg) {
 	while (1) {
-		while (isspace(*s)) {
+		while (isspace((int)*s)) {
 			if (*s == 0)
 				return "";
 			s++;
@@ -985,7 +985,7 @@ const char *STR_FindArg(const char *s, int arg) {
 		if (arg < 0) {
 			return s;
 		}
-		while (isspace(*s) == false) {
+		while (isspace((int)*s) == false) {
 			if (*s == 0)
 				return "";
 			s++;
